Moved the channel mappings and attributes to the config section

+++ apps/px4io/px4io.h
@@ -75,7 +75,7 @@ struct sys_state_s {
 	/**
 	 * Remote control input(s) channel mappings
 	 */
-	uint16_t	rc_map[4];
+	uint8_t	rc_map[4];
 
 	/**
 	 * Remote control channel attributes 
@@ -85,7 +85,7 @@ struct sys_state_s {
 	uint16_t	rc_max[4];
 	uint16_t	rc_rev[4];
 	uint16_t	rc_dz[4];
-	
+
 	/**
 	 * Data from the remote control input(s)
 	 */
